refactor(query): rename queryStreaming.ts -> query.ts
