+++ src/components/deck/DeckMentionsColumn.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed } from 'vue'
-import { createQuerySubscription } from '@/adapters/misskey/queryStreaming'
+import { createQuerySubscription } from '@/adapters/misskey/query'
 import type { NormalizedNote } from '@/adapters/types'
 import ColumnEmptyState from '@/components/common/ColumnEmptyState.vue'
 import LoadingSpinner from '@/components/common/LoadingSpinner.vue'
